moved get_start_datetime to convert_session
