+++ src/jadhav_lab_to_nwb/olson_2024/__init__.py
@@ -5,4 +5,4 @@ from .olson_2024_spike_gadgets_recording_interface import Olson2024SpikeGadgetsR
 from .olson_2024_sorting_interface import Olson2024SortingInterface
 from .olson_2024_spike_gadgets_lfp_interface import Olson2024SpikeGadgetsLFPInterface
 from .olson_2024_epoch_interface import Olson2024EpochInterface
-from .olson_2024_nwbconverter import Olson2024NWBConverter, get_start_datetime
+from .olson_2024_nwbconverter import Olson2024NWBConverter
