+++ src/jadhav_lab_to_nwb/olson_2024/olson_2024_nwbconverter.py
@@ -26,21 +26,3 @@ class Olson2024NWBConverter(NWBConverter):
         Behavior=Olson2024BehaviorInterface,
         Epoch=Olson2024EpochInterface,
     )
-
-
-def get_start_datetime(epoch_folder_name: str) -> datetime:
-    """Get the start datetime of the epoch from the folder name.
-
-    Parameters
-    ----------
-    epoch_folder_path : pathlib.Path
-        The path to the epoch folder.
-
-    Returns
-    -------
-    datetime.datetime
-        The start datetime of the epoch.
-    """
-    split_name = epoch_folder_name.split("_")
-    start_datetime = datetime.strptime(split_name[-2] + "_" + split_name[-1], "%Y%m%d_%H%M%S")
-    return start_datetime
